refactor(cli, controllers): naming + file organization
Two small cleanups from the deep audit:

1. helmrelease.dispatchRendered → emitRenderedChildren
   Kustomization and HelmRelease controllers both have a method that
   parses rendered docs and lands them in the store. KS calls it
   emitRenderedChildren; HR called it dispatchRendered. Same role,
   different names — readers had to learn both. Renamed HR's to match.

2. internal/cli/helpers.go → split by concern
   helpers.go held two unrelated bundles: image-extraction helpers
   (collectImages, emitImageList) and diff-orchestrator helpers
   (runDiffOrchestrators, joinRunErrors, gatherArtifacts, etc.).
   Image helpers move to images.go. Diff helpers fold into diff.go
   alongside the diff command builder (where they're called from).
   helpers.go shrinks to firstArg — the one genuine general utility.

Net: same code, organized by what it does.

Audit-D.
